+++ lib/datadog/core/telemetry/http/transport.rb
